Domínio do lojista: FKs compostas por tenant, trigger de estoque, campos novos e limpeza do legado
- P8: converte ~34 FKs entre tabelas de dominio para FOREIGN KEY (id_tenant, id_x) —
  FK simples nao aplica RLS na checagem de integridade referencial, permitindo um
  tenant referenciar registros de outro; achado confirmado em teste ao vivo e agora
  bloqueado pelo banco.
- Trigger trg_produto_movimento_detalhe_estoque (fn_atualiza_estoque_movimento)
  mantem produto_estoque.qtd_estoque via INSERT/UPDATE/DELETE em
  produto_movimento_detalhe, que deixa de ser imutavel (produto_movimento_mestre
  continua). Substitui a antiga SP_ATUALIZA_QUANTIDADE_ESTOQUE.
- Bug do bootstrap corrigido: niner_owner ganha GRANT CREATE ON SCHEMA public
  (sem isso o Flyway falhava em V013 fora do caminho de teste).
- Reverte regressao em produto_barra (sku/ean, Q7) introduzida por edicao anterior.
- Novos campos: cliente (categoria obrigatoria, data_nascimento/genero condicionados
  a pessoa fisica, redes sociais), funcionario.telefone, empresa.cfg_nome_etiqueta,
  cfg_geral (variantes linha/coluna, sem moeda_devolucao), produto_imagem (galeria).
  cliente_dados_pessoais_ck, funcionario_cpf_uk ajustada, venda sem id_funcionario.
- Padronizacao de tipos (id_tenant smallint, demais PKs integer) e remocao de
  ON DELETE CASCADE em todo o dominio, decisoes ja registradas nesta sessao.
- Documentacao (db/migration/README.md, spec-driven-erp-varejo.md, docs/PROGRESSO.md)
  sincronizada com cada mudanca; banco recreado e testado do zero varias vezes
  (isolamento entre tenants, trigger, FKs compostas).
- Remove os dumps legados db/*.txt (Firebird) — schema ja totalmente migrado e
  superado pelas migrations V013-V024.

+++ db/migration/V010__plataforma_staff.sql
@@ -19,7 +19,7 @@ COMMENT ON TABLE plataforma.staff IS 'Staff da plataforma (SUPER_ADMIN/SUPORTE/F
 CREATE TABLE plataforma.impersonacao_log (
   id           integer     GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   id_staff     integer     NOT NULL REFERENCES plataforma.staff (id_staff),
-  id_tenant    integer     NOT NULL REFERENCES plataforma.tenant (id_tenant),
+  id_tenant    smallint    NOT NULL REFERENCES plataforma.tenant (id_tenant),
   iniciado_em  timestamptz NOT NULL DEFAULT now(),
   encerrado_em timestamptz,
   motivo       text
